server, ntdll: Also wait on the queue fd when waiting for driver events.
Normally the server handles this, by polling on the fd during its main loop.
The problem there is that the server only polls when the thread is waiting
for messages. We want to replicate that behaviour, otherwise wineserver spins
forever trying to wake up a thread that just doesn't care.

With this patch, I'm finally able to interact with winecfg. Next step is to
fix the 'drives' tab.